Add normal hasIntersection and create Permutations class
